refactor(targets): introduce TargetAdapter dataclass + decompose _common.py (#219)
Mirror the aggregate/_adapter.py:SourceAdapter pattern on the target
side. Each target builder now declares a module-level
ADAPTER = TargetAdapter(...) and delegates to a generic
targets/_driver.py:build(adapter, project) that consumes the adapter
metadata and runs the shared pipeline (read aggregated NCs → combine →
optional NN-fill → atomic CF-1.6 write; year-chunked builders get
fingerprint-cached skip + orphan-prune + stitch on top).

The monolithic targets/_common.py (1,557 LOC) decomposes by concern
into _shims (SourceShim + indexers), _io (read_aggregated_source +
reindex_to_*_start + compute_hru_* + parse_period + check_hru_coords),
_combine (multi_source_nanminmax + build_n_sources_attrs), _writers
(write_target_nc + write_bounds_target), and _intermediates
(iter_period_years + fingerprint cache + prune + stitch). _common.py
becomes a thin back-compat re-export so every existing test import
continues to work unchanged.

SCA's per-year action stays as a module-level _build_year function
that owns load + write together so the test fixture's
monkeypatch.setattr(sca, "_build_year", ...) continues to no-op the
year loop and exercise the stitch-failure path. SWE uses the generic
year-chunked driver path. Single-shot builders (runoff, AET, recharge,
soil moisture) use the generic single-shot driver path; SOM declares
two adapters (monthly + annual) and drives them sequentially because
multi-variant output is out of scope for the driver contract.

Behavior changes: none. All 1,209 existing tests pass without
modification.
