+++ src/nhf_spatial_targets/targets/_common.py
@@ -1,1557 +1,110 @@
-"""Shared multi-source-minmax target machinery.
+"""Back-compat re-export of the decomposed target-builder helpers.
 
-Used by ``targets/run.py`` (and, in future, ``targets/aet.py``) to:
+This module used to be the 1,557-line single file containing every shared
+target-builder helper. It has been broken up by concern (PR #219) into:
 
-- read per-year aggregated source NCs as a single lazy xarray DataArray;
-- canonicalize cross-source monthly time coordinates onto a master
-  month-start index;
-- combine sources with NaN-aware reductions and emit a per-cell finite-
-  source-count diagnostic;
-- compute HRU area and centroids in both equal-area and lat/lon CRSes;
-- write CF-1.6 compliant NetCDF output atomically.
+- ``_shims``         — :class:`SourceShim` and the by-key / by-label /
+                       validate_source_units indexers.
+- ``_io``            — :func:`read_aggregated_source`,
+                       :func:`reindex_to_month_start` /
+                       :func:`reindex_to_day_start`,
+                       :func:`compute_hru_*`, :func:`parse_period`,
+                       :func:`check_hru_coords`.
+- ``_combine``       — :func:`multi_source_nanminmax`,
+                       :func:`build_n_sources_attrs`.
+- ``_writers``       — :func:`write_target_nc`,
+                       :func:`write_bounds_target`.
+- ``_intermediates`` — :func:`iter_period_years`,
+                       :func:`target_config_fingerprint`,
+                       :func:`code_version_fingerprint`,
+                       :func:`should_skip_year_build`,
+                       :func:`prune_orphan_year_intermediates`,
+                       :func:`stitch_year_chunks_to_target`, and the
+                       ``INTERMEDIATE_*_ATTR`` constants.
 
-Linear unit conversions live in the per-target module (e.g. ``run.py``
-keeps ``mm_per_month_to_cfs``); this module is unit-agnostic.
+This module re-exports every public (and a handful of private) symbol
+so existing imports continue to work unchanged::
+
+    from nhf_spatial_targets.targets._common import read_aggregated_source
+    from nhf_spatial_targets.targets._common import (
+        SourceShim, multi_source_nanminmax, write_bounds_target,
+    )
+
+Prefer the focused submodules for new code; this aggregator exists
+purely to keep PR #219 a refactor (no test changes).
 """
 
 from __future__ import annotations
 
-import hashlib
-import json
-import logging
-import re
-from collections.abc import Callable
-from dataclasses import dataclass
-from pathlib import Path
-
-import numpy as np
-import pandas as pd
-import xarray as xr
-
-from nhf_spatial_targets.workspace import Project
-
-logger = logging.getLogger(__name__)
-
-
-@dataclass(frozen=True)
-class SourceShim:
-    """Per-source contract for a multi-source target builder.
-
-    Co-locates the three facts each target needs about a contributing
-    source: the catalog key it loads from, the variable name in the
-    aggregated NC, the human-readable label for the output NC's global
-    ``source`` attr, and the per-source unit-shim function that brings the
-    native variable into the target's common intermediate units (e.g.
-    mm/month for runoff and AET).
-
-    Target modules declare a tuple of :class:`SourceShim` instances as
-    their ``SHIMS`` constant and call :func:`shims_by_key` to look one up
-    by catalog key inside their build loop. Adding a future source is a
-    single edit — there is no parallel-dict drift surface.
-
-    Attributes
-    ----------
-    source_key
-        On-disk storage key. Used as the subdirectory name under
-        ``<project>/data/aggregated/`` and as the prefix of the per-year
-        NC filename. Matches the catalog source key in the common case,
-        but may be a synthetic key (e.g. ``"era5_land_sd"``) when one
-        upstream source produces multiple aggregated cadences in
-        separate subdirs — see ``aggregate/era5_land.py``'s
-        ``ADAPTER_SD``.
-    aggregated_var
-        Variable name to extract from the aggregated NC (e.g. ``"et"``).
-    description
-        Human-readable label for the output NC's global ``source`` attr.
-    to_common_units
-        Callable that converts the source's native variable to the
-        target's common intermediate units. For multi_source_minmax
-        targets the common intermediate is usually mm/month; the per-
-        target unit chain then applies a final linear conversion (e.g.
-        mm/month → cfs for runoff, mm/month → inches/day for AET).
-    config_label
-        Optional user-facing alias used in the project config's
-        ``<target>.sources`` list. Defaults to ``source_key``. Set this
-        when the storage key is a synthetic disambiguation (e.g.
-        ``"era5_land_sd"``) but the config should keep the canonical
-        catalog name (``"era5_land"``) — this keeps target builders
-        free of any parallel label→storage dict that could drift from
-        the SHIMS registry.
-    catalog_source_key
-        Optional ``catalog/sources.yml`` key for the source. Defaults
-        to ``config_label`` if set, else ``source_key``. Used by
-        :func:`validate_source_units` to look up the catalog
-        ``cf_units`` for ``aggregated_var``. Set this only when the
-        catalog key is genuinely independent of both the on-disk
-        storage key and the user-facing config alias — e.g. a shim
-        named ``foo_v2`` with ``config_label="legacy_foo"`` (a config
-        alias) but catalog data still under ``foo_v2``.
-    expected_cf_units
-        Optional CF-style units string this shim was written against.
-        When set, :func:`validate_source_units` (called by every
-        target builder at startup) asserts the catalog's ``cf_units``
-        for ``aggregated_var`` matches exactly. Catches drift between
-        the catalog and the shim's hardcoded conversion factor — e.g.
-        a post-hoc cf_units correction that the shim hasn't been
-        updated for. ``None`` opts out of the check (used for
-        synthetic source keys whose aggregated variable is not the
-        same name as anything in the catalog).
-    """
-
-    source_key: str
-    aggregated_var: str
-    description: str
-    to_common_units: Callable[[xr.DataArray], xr.DataArray]
-    config_label: str | None = None
-    catalog_source_key: str | None = None
-    expected_cf_units: str | None = None
-
-
-def shims_by_key(shims: "tuple[SourceShim, ...]") -> "dict[str, SourceShim]":
-    """Index a ``SHIMS`` tuple by ``source_key`` for lookup at build time.
-
-    Raises ``ValueError`` if two shims share the same ``source_key`` —
-    that would silently shadow one entry, which is exactly the kind of
-    drift this refactor prevents.
-    """
-    out: dict[str, SourceShim] = {}
-    for shim in shims:
-        if shim.source_key in out:
-            raise ValueError(
-                f"Duplicate SourceShim.source_key={shim.source_key!r} in "
-                f"target SHIMS registry. Each source may appear at most once."
-            )
-        out[shim.source_key] = shim
-    return out
-
-
-def validate_source_units(
-    shims: "tuple[SourceShim, ...]",
-    sources: "list[str] | tuple[str, ...]",
-) -> None:
-    """Assert each requested source's catalog ``cf_units`` matches the shim.
-
-    Resolves each entry in ``sources`` (a list of config labels, e.g.
-    ``["era5_land", "gldas_noah_v21_monthly"]`` from the project config)
-    to its :class:`SourceShim`, then looks up the catalog ``cf_units``
-    for the shim's ``aggregated_var`` under the shim's catalog source
-    key (``catalog_source_key`` if set, else ``config_label``, else
-    ``source_key``) and raises if the strings differ.
-
-    Called at the top of every target builder's ``build()`` so a
-    post-hoc catalog correction (PR #68 caught four such corrections)
-    fails loud at startup rather than silently producing values off by
-    the missed conversion factor. Per-source unit shims encode an
-    assumption — e.g. ``gldas_to_mm_per_month`` assumes ``cf_units:
-    "kg m-2"`` so it can apply ``× 8 × days_in_month``. If the catalog
-    is corrected to a different string (or the shim's conversion is
-    updated without touching ``expected_cf_units``) this validator
-    raises with both strings so the operator can decide which side
-    needs to follow the other.
-
-    Shims with ``expected_cf_units=None`` opt out (e.g. legacy stubs
-    pending unit harmonisation). Unknown source labels raise too —
-    catching a project-config typo at startup is the same drift class
-    this validator guards.
-
-    Parameters
-    ----------
-    shims
-        The target's ``SHIMS`` tuple.
-    sources
-        Project-config-side source labels for which the build is
-        about to read data. Typically ``runoff_cfg["sources"]``,
-        ``aet_cfg["sources"]``, etc.
-
-    Raises
-    ------
-    ValueError
-        Catalog ``cf_units`` differs from ``shim.expected_cf_units``
-        for any requested source, or a requested source label has no
-        matching shim.
-    """
-    from nhf_spatial_targets import catalog as cat
-
-    by_label = shims_by_config_label(shims)
-    for label in sources:
-        if label not in by_label:
-            raise ValueError(
-                f"validate_source_units: no matching SourceShim for source "
-                f"{label!r}. Known labels: {sorted(by_label)}."
-            )
-        shim = by_label[label]
-        if shim.expected_cf_units is None:
-            continue
-        catalog_key = shim.catalog_source_key or shim.config_label or shim.source_key
-        try:
-            actual = cat.source_var_cf_units(catalog_key, shim.aggregated_var)
-        except KeyError as exc:
-            raise ValueError(
-                f"validate_source_units: cannot resolve cf_units for "
-                f"{catalog_key!r}/{shim.aggregated_var!r}: {exc}. Either "
-                f"add cf_units to catalog/sources.yml or set "
-                f"expected_cf_units=None on the shim to opt out."
-            ) from exc
-        if actual != shim.expected_cf_units:
-            raise ValueError(
-                f"Catalog cf_units drift for {catalog_key!r}/"
-                f"{shim.aggregated_var!r}: catalog has {actual!r}, "
-                f"shim {shim.source_key!r} expects "
-                f"{shim.expected_cf_units!r}. Update the shim if the "
-                f"units changed intentionally, or correct the catalog."
-            )
-
-
-def shims_by_config_label(
-    shims: "tuple[SourceShim, ...]",
-) -> "dict[str, SourceShim]":
-    """Index a ``SHIMS`` tuple by ``config_label`` (defaulted to ``source_key``).
-
-    Used by target builders to resolve a user-facing source name from
-    the project config (e.g. ``"era5_land"``) to the SourceShim that
-    encodes its storage key, aggregated variable, and unit shim. Raises
-    ``ValueError`` on duplicate labels — two shims wanting the same
-    config label is the same drift class :func:`shims_by_key` guards.
-    """
-    out: dict[str, SourceShim] = {}
-    for shim in shims:
-        label = shim.config_label or shim.source_key
-        if label in out:
-            raise ValueError(
-                f"Duplicate SourceShim config_label={label!r} in target "
-                f"SHIMS registry. Each label may appear at most once "
-                f"(set distinct config_label values to disambiguate)."
-            )
-        out[label] = shim
-    return out
-
-
-_AGG_READ_CHUNK_BYTES = 256 * 1024 * 1024
-
-
-def _read_chunk_hru(
-    n_time: int, itemsize: int, target_bytes: int = _AGG_READ_CHUNK_BYTES
-) -> int:
-    """HRU-dim dask chunk length for reading aggregated NCs (#165 ST3).
-
-    Reads use full-time chunks (``{"time": -1}``) so a file's on-disk time
-    chunk is never split (which would re-inflate the same compressed chunk on
-    every time slab). The HRU dim is chunked to ~``target_bytes`` per chunk to
-    keep per-chunk memory and the dask graph reasonable — deliberately larger
-    than the on-disk ~1 MiB HRU chunk, trading chunk granularity for a smaller
-    graph on a full-source read.
-    """
-    return max(1, target_bytes // (max(n_time, 1) * itemsize))
-
-
-def read_aggregated_source(
-    project: Project,
-    source_key: str,
-    var: str,
-    period: tuple[str, str],
-    chunks: dict | None = None,
-) -> xr.DataArray:
-    """Open per-year aggregated NCs for one source and return one variable.
-
-    Reads ``<project.aggregated_dir()>/<source_key>/<source_key>_*_agg.nc``
-    via ``xr.open_mfdataset`` (lazy / dask-backed), slices to the requested
-    period, and returns the requested variable as a DataArray.
-
-    The returned DataArray is lazy: it holds open file handles via the dask
-    graph. The caller is responsible for consuming it (e.g. via
-    ``.compute()``, ``.load()``, or a streaming ``to_netcdf``) before the
-    underlying dataset goes out of scope. Do **not** call ``.close()`` on
-    the returned DataArray directly — it does not own the dataset handle.
-
-    The HRU dim name in the aggregated NCs matches ``project.id_col`` (e.g.
-    ``nhm_id``).
-
-    Parameters
-    ----------
-    project
-        Loaded :class:`~nhf_spatial_targets.workspace.Project`.
-    source_key
-        Catalog key (e.g. ``"era5_land"``).
-    var
-        Variable name to extract from the aggregated dataset.
-    period
-        ``(start_iso, end_iso)`` tuple, both inclusive (e.g.
-        ``("2000-01-01", "2010-12-31")``).
-    chunks
-        Forwarded to ``xr.open_mfdataset``. When ``None``, defaults to
-        ``{"time": -1, project.id_col: <chunk_hru>}`` — full time per chunk,
-        HRU dim chunked to ~256 MiB — to align reads with the columnar
-        on-disk layout written by the aggregator (#165 ST3). Pass an explicit
-        dict to override.
-
-    Raises
-    ------
-    FileNotFoundError
-        If the source's aggregated directory contains no per-year NCs.
-    ValueError
-        If the requested period falls entirely outside the source's
-        per-year coverage.
-    """
-    agg_dir = project.aggregated_dir() / source_key
-    pattern = f"{source_key}_*_agg.nc"
-    paths = sorted(agg_dir.glob(pattern))
-    if not paths:
-        raise FileNotFoundError(
-            f"No aggregated NC files found for source '{source_key}' under "
-            f"{agg_dir} (pattern: {pattern}). Run "
-            f"'pixi run nhf-targets agg {source_key.replace('_', '-')} "
-            f"--project-dir {project.workdir}' first."
-        )
-
-    if chunks is None:
-        # Align dask chunks to the new columnar on-disk layout (#165 ST3):
-        # full time per chunk so a file's on-disk time chunk is never split
-        # (splitting would force the compressor to re-inflate the same chunk
-        # on every time slab — ~30x amplification on a daily source), with the
-        # HRU dim chunked to ~256 MiB. Works for both the new chunked NCs and
-        # legacy contiguous NCs (the latter read as bounded strided slabs, not
-        # a whole-array load). open_mfdataset chunks per file along the concat
-        # dim, so ``{"time": -1}`` yields one time chunk per file — size the
-        # HRU chunk from one file's time length (probed cheaply from the first
-        # file) so each per-file chunk lands near the byte budget.
-        with xr.open_dataset(paths[0], engine="netcdf4") as probe:
-            n_time = int(probe.sizes.get("time", 1))
-            itemsize = probe[var].dtype.itemsize if var in probe.data_vars else 8
-        chunks = {"time": -1, project.id_col: _read_chunk_hru(n_time, itemsize)}
-
-    ds = xr.open_mfdataset(
-        [str(p) for p in paths],
-        combine="by_coords",
-        join="outer",
-        chunks=chunks,
-        engine="netcdf4",
-    )
-    if var not in ds:
-        available = sorted(ds.data_vars)
-        ds.close()
-        raise KeyError(
-            f"Variable '{var}' not found in aggregated NCs for source "
-            f"'{source_key}'. Available variables: {available}."
-        )
-    # Canonical row order: HRU dim ascending by id_col. Emission-time
-    # enforcement landed with issue #93, so per-year NCs aggregated after
-    # that change arrive here already sorted (this call is a no-op for
-    # them). For pre-#93 NCs already on disk — where gdptools wrote rows
-    # in VPU-grouped batch order — this defensive sort keeps positional
-    # checks against the fabric correct without forcing a re-aggregate.
-    ds = ds.sortby(project.id_col)
-    # Substitute netCDF default-fill cells (NC_FILL_DOUBLE etc.) with NaN.
-    # Aggregated NCs written before issue #204's fix declare _FillValue=NaN
-    # but hold the ~1e36 sentinel in cells gdptools couldn't fill (e.g. OR
-    # HRUs outside MWBM CONUS extent). xarray's decode masks via value
-    # equality and NaN never matches, so the sentinel survives and poisons
-    # downstream np.fmax. Mirrors the write-side fix in
-    # aggregate/_driver.py:_atomic_write_netcdf — keeping both means
-    # existing on-disk NCs decode correctly without a rewrite while fresh
-    # writes are self-consistent on disk.
-    from nhf_spatial_targets.io_nc import mask_netcdf_default_fills
-
-    ds = mask_netcdf_default_fills(ds)
-    sliced = ds[var].sel(time=slice(period[0], period[1]))
-    if sliced.sizes.get("time", 0) == 0:
-        # Years parsed from sorted filenames -- avoids triggering dask compute
-        # on a possibly large coordinate array, and is robust to time-coord
-        # corruption inside the NCs.
-        first_year = paths[0].name.rsplit("_", 2)[-2]
-        last_year = paths[-1].name.rsplit("_", 2)[-2]
-        ds.close()
-        raise ValueError(
-            f"Requested period {period[0]} .. {period[1]} is entirely "
-            f"outside source coverage for '{source_key}' "
-            f"(years {first_year} .. {last_year})."
-        )
-    logger.info(
-        "Loaded %s/%s: %d months from %d per-year NCs",
-        source_key,
-        var,
-        sliced.sizes["time"],
-        len(paths),
-    )
-    return sliced
-
-
-def reindex_to_month_start(
-    da: xr.DataArray, master_index: pd.DatetimeIndex
-) -> xr.DataArray:
-    """Reindex a monthly DataArray onto a master ``freq="MS"`` index.
-
-    Source timestamps may be end-of-month (ERA5-Land), start-of-month (GLDAS,
-    MWBM), or mid-month (MERRA-2 etc.). All three convey "which calendar
-    month" unambiguously. This helper converts the source's time coordinate
-    via ``dt.to_period("M").dt.to_timestamp()`` (yielding the month-start),
-    then reindexes onto ``master_index``.
-
-    Months in ``master_index`` that the source does not cover come back as
-    NaN — this is what gives the runoff target its period-union semantics:
-    a source that ends in 2020 but is asked through 2024 simply contributes
-    nothing for the post-2020 cells.
-
-    Parameters
-    ----------
-    da
-        Monthly DataArray to reindex.
-    master_index
-        Target index. Must be ``DatetimeIndex`` with ``freq="MS"``.
-    """
-    if not isinstance(master_index, pd.DatetimeIndex):
-        raise TypeError(
-            f"master_index must be a pandas.DatetimeIndex, got "
-            f"{type(master_index).__name__}"
-        )
-    if master_index.freqstr != "MS":
-        raise ValueError(
-            f"master_index must have freq='MS' (month-start); got "
-            f"freq={master_index.freqstr!r}. Build it with "
-            f"pd.date_range(start, end, freq='MS')."
-        )
-    ms_times = pd.DatetimeIndex(da.time.values).to_period("M").to_timestamp()
-    canon = da.assign_coords(time=ms_times)
-    return canon.reindex(time=master_index)
-
-
-def reindex_to_day_start(
-    da: xr.DataArray, master_index: pd.DatetimeIndex
-) -> xr.DataArray:
-    """Reindex a daily DataArray onto a master ``freq="D"`` index.
-
-    Source timestamps may be midnight (SNODAS, ERA5-Land daily, Margulis)
-    or noon (Daymet — calendar-day mean). Both convey "which calendar
-    day" unambiguously; this helper normalises by stripping the time of
-    day (``.dt.floor("D")``) before reindexing.
-
-    Days in ``master_index`` that the source does not cover come back as
-    NaN — this is what gives the SWE target its period-union semantics:
-    a source whose record ends in 2020 but is asked through 2024 simply
-    contributes nothing for the post-2020 cells.
-
-    Requires ``da.time`` to be decoded as ``datetime64[ns]`` (the xarray
-    default for ``proleptic_gregorian`` calendars, which is what every
-    NC the pipeline writes uses — see ``fetch/consolidate.py``). Sources
-    decoded as ``cftime`` objects (non-standard calendars like
-    ``noleap`` / ``360_day``) raise ``TypeError`` rather than silently
-    falling back to a ``DatetimeIndex`` conversion that loses the
-    calendar.
-    """
-    if not isinstance(master_index, pd.DatetimeIndex):
-        raise TypeError(
-            f"master_index must be a pandas.DatetimeIndex, got "
-            f"{type(master_index).__name__}"
-        )
-    if master_index.freqstr != "D":
-        raise ValueError(
-            f"master_index must have freq='D' (daily); got "
-            f"freq={master_index.freqstr!r}. Build it with "
-            f"pd.date_range(start, end, freq='D')."
-        )
-    if not np.issubdtype(da.time.dtype, np.datetime64):
-        raise TypeError(
-            f"reindex_to_day_start expects datetime64-decoded time, got "
-            f"dtype={da.time.dtype!r}. The pipeline writes every NC with "
-            f"calendar='proleptic_gregorian' so xarray should decode to "
-            f"datetime64[ns]; a cftime-decoded source indicates either a "
-            f"non-standard upstream calendar or decode_cf=False during "
-            f"open. Re-open with the default decoder or report as a "
-            f"consolidator bug."
-        )
-    day_times = pd.DatetimeIndex(da.time.values).floor("D")
-    canon = da.assign_coords(time=day_times)
-    return canon.reindex(time=master_index)
-
-
-def multi_source_nanminmax(
-    sources: dict[str, xr.DataArray],
-) -> tuple[xr.DataArray, xr.DataArray, xr.DataArray]:
-    """NaN-aware per-cell min, max, and finite-source count.
-
-    All input DataArrays must share dims and coords (typically
-    ``(time, id_col)``). They are stacked on a new ``source`` dim and
-    reduced with ``skipna=True``.
-
-    A bound is defined whenever ≥1 source is finite at that cell; the result
-    is NaN only when *every* source is NaN there, which is exactly when
-    ``n_sources == 0``.
-
-    Parameters
-    ----------
-    sources
-        Mapping from source key to per-source DataArray.
-
-    Returns
-    -------
-    lower, upper, n_sources
-        ``(time, id_col)`` arrays. ``n_sources`` is int8 with values in
-        ``[0, len(sources)]``; ``lower`` / ``upper`` preserve the input
-        dtype (typically float32).
-
-    Raises
-    ------
-    ValueError
-        If any two sources have different HRU coords (different fabrics).
-    """
-    keys = list(sources.keys())
-    if not keys:
-        raise ValueError("multi_source_nanminmax: empty sources dict")
-    ref = sources[keys[0]]
-    hru_dim = next(d for d in ref.dims if d != "time")
-    for k in keys[1:]:
-        other = sources[k]
-        if not other[hru_dim].equals(ref[hru_dim]):
-            raise ValueError(
-                f"HRU coords differ between sources '{keys[0]}' and '{k}'. "
-                "All sources must be aggregated to the same fabric."
-            )
-
-    stacked = xr.concat([sources[k] for k in keys], dim=xr.Variable("source", keys))
-    lower = stacked.min(dim="source", skipna=True)
-    upper = stacked.max(dim="source", skipna=True)
-    n_sources = stacked.notnull().sum(dim="source").astype(np.int8)
-    return lower, upper, n_sources
-
-
-def _load_and_reproject_fabric(project: Project) -> "tuple[object, str]":
-    """Load the fabric, validate the id_col, reproject to ``area_crs``.
-
-    Shared expensive setup for :func:`compute_hru_areas`,
-    :func:`compute_hru_centroids`, and the combined
-    :func:`compute_hru_area_and_centroids`. The fabric IO and the CRS
-    reprojection are the two non-trivial costs at fabric scale (~361k
-    polygons on gfv2); per-column derivations (area, centroids) are
-    near-free relative to those.
-
-    Returns
-    -------
-    (gdf_eq, id_col)
-        Reprojected GeoDataFrame in ``project.area_crs`` and the validated
-        HRU id column name.
-    """
-    import geopandas as gpd
-
-    fabric_path = Path(project.config["fabric"]["path"])
-    if fabric_path.suffix.lower() in (".parquet", ".geoparquet"):
-        gdf = gpd.read_parquet(fabric_path)
-    else:
-        gdf = gpd.read_file(fabric_path)
-    id_col = project.id_col
-    if id_col not in gdf.columns:
-        raise ValueError(
-            f"Column '{id_col}' not found in fabric {fabric_path}. "
-            f"Available: {list(gdf.columns)}"
-        )
-    if not gdf[id_col].is_unique:
-        n_dupes = (gdf[id_col].value_counts() > 1).sum()
-        raise ValueError(
-            f"Fabric column '{id_col}' has {n_dupes} duplicate values "
-            f"in {fabric_path}. Each HRU must have a unique ID."
-        )
-
-    gdf_eq = gdf.to_crs(project.area_crs)
-    return gdf_eq, id_col
-
-
-def compute_hru_centroids(project: Project) -> "pd.DataFrame":
-    """Compute per-HRU centroid coords from the fabric.
-
-    Reprojects to ``project.area_crs`` (e.g. EPSG:5070 for CONUS) for
-    equal-area centroids, then reprojects centroids to EPSG:4326 for
-    ancillary lat/lon. Use this in target builders that do **not** need
-    per-HRU area (AET, recharge, soil moisture, SCA, SWE) — it skips the
-    `gdf.geometry.area` call that :func:`compute_hru_area_and_centroids`
-    performs.
-
-    Returns
-    -------
-    pandas.DataFrame
-        Indexed by ``project.id_col``. Columns: ``centroid_x``,
-        ``centroid_y`` (in ``area_crs``), ``centroid_lat``,
-        ``centroid_lon`` (EPSG:4326).
-    """
-    gdf_eq, id_col = _load_and_reproject_fabric(project)
-    centroids_eq = gdf_eq.geometry.centroid
-    centroids_ll = centroids_eq.to_crs("EPSG:4326")
-
-    df = gdf_eq[[id_col]].copy()
-    df["centroid_x"] = centroids_eq.x.astype(float)
-    df["centroid_y"] = centroids_eq.y.astype(float)
-    df["centroid_lon"] = centroids_ll.x.astype(float)
-    df["centroid_lat"] = centroids_ll.y.astype(float)
-    df = df.set_index(id_col).sort_index()
-    return df
-
-
-def compute_hru_areas(project: Project) -> "pd.DataFrame":
-    """Compute per-HRU area (m²) from the fabric.
-
-    Reprojects to ``project.area_crs`` (e.g. EPSG:5070 for CONUS) so the
-    area is computed in an equal-area projection. Always recomputes from
-    geometry (no fabric-column fallback) so the area cannot drift from
-    the geometry actually being processed.
-
-    Returns
-    -------
-    pandas.DataFrame
-        Indexed by ``project.id_col``. Single column ``area_m2``.
-    """
-    gdf_eq, id_col = _load_and_reproject_fabric(project)
-    df = gdf_eq[[id_col]].copy()
-    df["area_m2"] = gdf_eq.geometry.area.astype(float)
-    df = df.set_index(id_col).sort_index()
-    return df
-
-
-def compute_hru_area_and_centroids(project: Project) -> "pd.DataFrame":
-    """Compute per-HRU area (m²) and centroid coords in a single fabric pass.
-
-    Combined helper for builders that need both (runoff). Lighter
-    alternatives are :func:`compute_hru_centroids` (centroids only) and
-    :func:`compute_hru_areas` (area only) — prefer them when you don't
-    need the other column.
-
-    Always recomputes from geometry (no fabric-column fallback) so the
-    area cannot drift from the geometry actually being processed.
-
-    Returns
-    -------
-    pandas.DataFrame
-        Indexed by ``project.id_col``. Columns: ``area_m2``,
-        ``centroid_x``, ``centroid_y`` (in ``area_crs``),
-        ``centroid_lat``, ``centroid_lon`` (EPSG:4326).
-    """
-    gdf_eq, id_col = _load_and_reproject_fabric(project)
-    centroids_eq = gdf_eq.geometry.centroid
-    centroids_ll = centroids_eq.to_crs("EPSG:4326")
-
-    df = gdf_eq[[id_col]].copy()
-    df["area_m2"] = gdf_eq.geometry.area.astype(float)
-    df["centroid_x"] = centroids_eq.x.astype(float)
-    df["centroid_y"] = centroids_eq.y.astype(float)
-    df["centroid_lon"] = centroids_ll.x.astype(float)
-    df["centroid_lat"] = centroids_ll.y.astype(float)
-    df = df.set_index(id_col).sort_index()
-    return df
-
-
-def _target_encoding_without_chunks(
-    ds: xr.Dataset, var_dtype: dict[str, str]
-) -> dict[str, dict]:
-    """Target encoding (dtype/zlib/fill/time) minus per-HRU ``chunksizes``.
-
-    Only used by :func:`write_target_nc` when no ``sort_dim``/``id_col`` is
-    supplied, so the per-HRU chunk dimension is unknown and HDF5 auto-chunks.
-    Reuses ``io_nc``'s fill-value and time policy so this fallback stays
-    consistent with the chunked ``build_encoding`` path.
-    """
-    from nhf_spatial_targets.io_nc import _TIME_ENCODING, _fill_value_for
-
-    encoding: dict[str, dict] = {}
-    for name, dt in var_dtype.items():
-        dtype = np.dtype(dt)
-        enc: dict = {
-            "dtype": str(dtype),
-            "zlib": True,
-            "complevel": 4,
-            "_FillValue": _fill_value_for(dtype),
-            # Explicit shuffle: netCDF4 defaults it True under zlib (see
-            # io_nc.build_encoding), so omitting it for floats would enable it.
-            "shuffle": bool(np.issubdtype(dtype, np.integer)),
-        }
-        encoding[name] = enc
-    for tvar in ("time", "time_bnds"):
-        if tvar in ds.variables:
-            encoding[tvar] = dict(_TIME_ENCODING)
-    # time_bnds is a CF boundary variable and must not carry _FillValue
-    # (CF §7.1); mirror io_nc.build_encoding and pin it off.
-    if "time_bnds" in ds.variables:
-        encoding["time_bnds"]["_FillValue"] = None
-    return encoding
-
-
-def write_target_nc(
-    ds: xr.Dataset,
-    output_path: Path,
-    title: str,
-    extra_global_attrs: dict | None = None,
-    sort_dim: str | None = None,
-) -> None:
-    """Write a target Dataset to NetCDF atomically with CF-1.6 metadata.
-
-    The Dataset is expected to already carry the data variables, ancillary
-    coordinates (``time_bnds``, ``centroid_lat``, ``centroid_lon``), and
-    per-variable attrs (``units``, ``long_name``, ``cell_methods``, etc.).
-    This helper sets the global ``Conventions`` / ``title`` / ``history`` /
-    ``software_version`` attrs, then delegates encoding to
-    :func:`io_nc.build_encoding` (``layer="target"``): float32+zlib bounds,
-    int8+zlib diagnostics, the pinned ``proleptic_gregorian`` time axis, and
-    per-HRU-time-series ``chunksizes`` so a single HRU's calibration read is
-    one ~1 MiB chunk (issue #165 ST2). The write goes through
-    :func:`io_nc.atomic_to_netcdf` (tempfile + rename) so a partial NetCDF
-    never lands at the final path. When ``sort_dim`` is omitted the HRU
-    chunk dim is unknown, so the same dtype/compression/time policy is applied
-    without per-HRU chunking.
-
-    When ``sort_dim`` is given, the Dataset is sorted ascending on that
-    dimension before write. Target builders pass ``project.id_col`` here
-    to enforce the canonical HRU row order at the emission boundary
-    (issue #93). Upstream helpers (``read_aggregated_source``,
-    ``compute_hru_area_and_centroids``) already produce sorted data; the
-    explicit sort here makes the invariant unmistakable at the file boundary.
-    """
-    from datetime import datetime, timezone
-
-    from nhf_spatial_targets import __version__
-
-    output_path = Path(output_path)
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-
-    ds = ds.copy()
-    if sort_dim is not None:
-        ds = ds.sortby(sort_dim)
-    ds.attrs.setdefault("Conventions", "CF-1.6")
-    ds.attrs["title"] = title
-    ds.attrs["history"] = (
-        f"{datetime.now(timezone.utc).isoformat()} created by "
-        f"nhf_spatial_targets v{__version__}"
-    )
-    ds.attrs.setdefault("institution", "USGS")
-    ds.attrs.setdefault("software_version", __version__)
-    if extra_global_attrs:
-        ds.attrs.update(extra_global_attrs)
-
-    # CF §5.6: anchor the spatial reference. Target bound vars carry
-    # grid_mapping="crs" (inherited from the aggregated sources), but no crs
-    # container flows through the multi-source combine, leaving the reference
-    # dangling. The HRU centroid coords (centroid_lat/lon) are EPSG:4326, so
-    # mint a 0-dim WGS84 latitude_longitude grid-mapping variable to match.
-    from pyproj import CRS as _CRS
-
-    _wgs84 = _CRS.from_epsg(4326)
-    _crs_attrs = dict(_wgs84.to_cf())
-    _crs_attrs.setdefault("crs_wkt", _wgs84.to_wkt())
-    ds["crs"] = xr.DataArray(np.int32(0), attrs=_crs_attrs)
-    for _v in ("lower_bound", "upper_bound", "n_sources", "nn_filled"):
-        if _v in ds.data_vars:
-            ds[_v].attrs["grid_mapping"] = "crs"
-    # CF §3: the HRU index coordinate is an identifier, not a measurement —
-    # label it and carry no units.
-    if sort_dim is not None and sort_dim in ds.variables:
-        ds[sort_dim].attrs.pop("units", None)
-        ds[sort_dim].attrs["long_name"] = "HRU Index"
-
-    from nhf_spatial_targets.io_nc import atomic_to_netcdf, build_encoding
-
-    # Pin the on-disk dtype for each known target var: float32 bounds, int8
-    # diagnostics. build_encoding derives _FillValue / shuffle from the dtype
-    # (NaN + no-shuffle for floats, no-fill + shuffle for the int8 diagnostics).
-    target_dtypes = {
-        v: "float32" for v in ("lower_bound", "upper_bound") if v in ds.data_vars
-    }
-    target_dtypes.update(
-        {v: "int8" for v in ("n_sources", "nn_filled") if v in ds.data_vars}
-    )
-
-    if sort_dim is not None:
-        encoding = build_encoding(
-            ds, layer="target", hru_dim=sort_dim, var_dtype=target_dtypes
-        )
-    else:
-        # No id_col known — production target builders always pass sort_dim, so
-        # this is only the bare ``write_target_nc(ds, out, title=...)`` path.
-        # Apply the same dtype/compression/time policy minus per-HRU chunking
-        # (HDF5 auto-chunks), since the HRU dim name is unavailable here.
-        encoding = _target_encoding_without_chunks(ds, target_dtypes)
-
-    atomic_to_netcdf(ds, output_path, encoding=encoding)
-    logger.info("Wrote %s (%.1f MB)", output_path, output_path.stat().st_size / 1e6)
-
-
-# ---------------------------------------------------------------------------
-# Shared target-builder helpers
-# ---------------------------------------------------------------------------
-
-
-def parse_period(period_str: str) -> tuple[str, str]:
-    """Parse 'YYYY-MM-DD/YYYY-MM-DD' (or 'YYYY/YYYY') into ``(start, end)``.
-
-    Used by every target builder to split the project config's
-    ``<target>.period`` (and ``<target>.normalize_period`` where present)
-    into the two endpoints needed to slice ``read_aggregated_source``'s
-    output.
-    """
-    if "/" not in period_str:
-        raise ValueError(
-            f"Invalid period {period_str!r}. Expected 'YYYY-MM-DD/YYYY-MM-DD'."
-        )
-    start, end = period_str.split("/", 1)
-    return start.strip(), end.strip()
-
-
-def check_hru_coords(
-    da: xr.DataArray,
-    fabric_hru_ids: np.ndarray,
-    id_col: str,
-    source_key: str,
-) -> None:
-    """Raise if the source DataArray's HRU dim disagrees with the fabric.
-
-    Both sides are expected to be sorted ascending by ``id_col`` upstream
-    (``compute_hru_*`` helpers and ``read_aggregated_source`` enforce
-    this). Three outcomes:
-
-    - Coords match exactly: returns ``None``.
-    - Coords have the SAME SET but a different order: raises with a
-      "canonical-sort invariant regression" message — the upstream
-      sort-on-emission contract (#93) has been broken somewhere.
-    - Coords have different SETS: raises a "re-aggregate this source
-      against the current fabric" message.
-
-    Called by target builders right after ``read_aggregated_source``
-    so any HRU misalignment is caught before silent broadcast/intersect
-    arithmetic poisons downstream values.
-    """
-    src_hru_ids = da[id_col].values
-    if np.array_equal(src_hru_ids, fabric_hru_ids):
-        return
-    same_set = len(src_hru_ids) == len(fabric_hru_ids) and np.array_equal(
-        np.sort(src_hru_ids), np.sort(fabric_hru_ids)
-    )
-    if same_set:
-        raise ValueError(
-            f"HRU coords for source '{source_key}' have the same set as the "
-            f"fabric ({len(fabric_hru_ids)} HRUs) but a different order. "
-            f"Both sides are expected to be sorted ascending by id_col="
-            f"'{id_col}' — this indicates a regression in the canonical-"
-            f"sort invariant in targets/_common.py."
-        )
-    raise ValueError(
-        f"HRU coords differ between fabric and source '{source_key}' as "
-        f"sets. Fabric has {len(fabric_hru_ids)} HRUs "
-        f"(first={fabric_hru_ids[0]}, last={fabric_hru_ids[-1]}); "
-        f"source has {len(src_hru_ids)} "
-        f"(first={src_hru_ids[0]}, last={src_hru_ids[-1]}). "
-        f"Re-aggregate '{source_key}' against the current fabric."
-    )
-
-
-def build_n_sources_attrs(
-    n_sources_count: int,
-    ancillary_coords: str = "centroid_lat centroid_lon",
-) -> dict:
-    """Build the per-variable attrs dict for an ``n_sources`` diagnostic var.
-
-    Parameters
-    ----------
-    n_sources_count
-        Number of source contributors (an integer ≥ 0 and ≤ 5). Determines
-        the length of the ``flag_values`` list and the matching
-        ``flag_meanings`` labels (``none one two three four five``,
-        truncated to ``n_sources_count + 1`` entries).
-    ancillary_coords
-        Space-separated list of ancillary coordinate variable names to
-        record under CF's ``coordinates`` attr. Defaults to the centroid
-        pair used by every target builder.
-    """
-    flag_labels = ["none", "one", "two", "three", "four", "five"]
-    if n_sources_count + 1 > len(flag_labels):
-        raise ValueError(
-            f"build_n_sources_attrs: n_sources_count={n_sources_count} exceeds "
-            f"the {len(flag_labels) - 1}-source label vocabulary."
-        )
-    return {
-        "units": "1",
-        "long_name": "number of finite source contributions",
-        # int8 to match the on-disk n_sources dtype (CF §3.5 requires
-        # flag_values dtype == parent variable dtype).
-        "flag_values": np.array(range(0, n_sources_count + 1), dtype="int8"),
-        "flag_meanings": " ".join(flag_labels[: n_sources_count + 1]),
-        "coordinates": ancillary_coords,
-    }
-
-
-def write_bounds_target(
-    *,
-    project: Project,
-    lower: xr.DataArray,
-    upper: xr.DataArray,
-    n_sources: xr.DataArray,
-    n_sources_count: int,
-    time_index: pd.DatetimeIndex,
-    time_offset_unit: object,
-    bounds_units: str,
-    bounds_long_name_kind: str,
-    cell_methods: str,
-    output_path: Path,
-    title: str,
-    nn_title: str,
-    extra_global_attrs: dict,
-    hru_meta: "pd.DataFrame",
-    nn_fill: bool,
-    nn_max_candidates: int,
-    id_col: str,
-) -> None:
-    """Assemble + write a bounds-target Dataset, with optional NN-fill companion.
-
-    Consolidates the assemble-and-write pipeline shared by every target
-    builder (runoff, AET, recharge, soil moisture): centroid coords,
-    ``time_bnds``, per-variable attrs (units / long_name / cell_methods /
-    coordinates), global attrs, atomic write via ``write_target_nc``, a
-    coverage-summary log line, and the optional ``nn_fill_bounds``
-    companion file.
-
-    Parameters
-    ----------
-    project
-        Loaded :class:`~nhf_spatial_targets.workspace.Project`.
-    lower, upper, n_sources
-        The three combined-source DataArrays from
-        :func:`multi_source_nanminmax`.
-    n_sources_count
-        Total number of source contributors (an int); drives the
-        ``n_sources`` diagnostic's ``flag_values`` length.
-    time_index
-        Master ``DatetimeIndex`` that ``lower`` / ``upper`` are aligned to.
-    time_offset_unit
-        Offset added to each ``time_index`` entry to form ``time_bnds``'s
-        upper edge (e.g. ``pd.offsets.MonthBegin(1)`` for monthly,
-        ``pd.offsets.YearBegin(1)`` for annual).
-    bounds_units
-        Units string for the lower/upper variable attrs (e.g. ``"cfs"``,
-        ``"inches/day"``, ``"1"``).
-    bounds_long_name_kind
-        Substituted into the ``long_name`` template: ``"lower bound of
-        {kind} (NaN-aware min across sources)"``. Examples: ``"monthly
-        runoff"``, ``"annual recharge"``.
-    cell_methods
-        CF ``cell_methods`` attr value for both bounds (e.g.
-        ``"time: sum"``, ``"time: mean"``).
-    output_path
-        Final NetCDF path for the unfilled target.
-    title
-        ``title`` global attr for the unfilled target.
-    nn_title
-        ``title`` global attr for the NN-filled companion (only used when
-        ``nn_fill`` is True).
-    extra_global_attrs
-        Per-target metadata (``source``, ``period``, ``fabric_sha256``,
-        etc.) — passed through to ``write_target_nc``.
-    hru_meta
-        DataFrame returned by ``compute_hru_centroids`` (or the combined
-        helper). Must contain ``centroid_lat``, ``centroid_lon``,
-        ``centroid_x``, ``centroid_y`` columns.
-    nn_fill
-        If True, additionally write ``<output>_nn_filled.nc`` via
-        :func:`nn_fill_bounds`.
-    nn_max_candidates
-        Forwarded to :func:`nn_fill_bounds`.
-    id_col
-        HRU id column name (e.g. ``"nhm_id"``); the dataset is sorted
-        ascending on this dim at emission per the #93 canonical-row-order
-        invariant.
-    """
-    # Avoid a circular-import by deferring this helper-internal import.
-    from nhf_spatial_targets.normalize.methods import nn_fill_bounds
-
-    lower.name = "lower_bound"
-    upper.name = "upper_bound"
-    n_sources.name = "n_sources"
-
-    time_bnds = xr.DataArray(
-        list(zip(time_index.values, (time_index + time_offset_unit).values)),
-        dims=("time", "nv"),
-        coords={"time": time_index.values},
-        name="time_bnds",
-    )
-    centroid_lat = xr.DataArray(
-        hru_meta["centroid_lat"].values,
-        dims=(id_col,),
-        coords={id_col: hru_meta.index.values},
-        attrs={
-            "units": "degrees_north",
-            "standard_name": "latitude",
-            "long_name": "HRU centroid latitude",
-        },
-    )
-    centroid_lon = xr.DataArray(
-        hru_meta["centroid_lon"].values,
-        dims=(id_col,),
-        coords={id_col: hru_meta.index.values},
-        attrs={
-            "units": "degrees_east",
-            "standard_name": "longitude",
-            "long_name": "HRU centroid longitude",
-        },
-    )
-
-    lower.attrs.update(
-        {
-            "units": bounds_units,
-            "long_name": (
-                f"lower bound of {bounds_long_name_kind} (NaN-aware min across sources)"
-            ),
-            "cell_methods": cell_methods,
-            "coordinates": "centroid_lat centroid_lon",
-        }
-    )
-    upper.attrs.update(
-        {
-            "units": bounds_units,
-            "long_name": (
-                f"upper bound of {bounds_long_name_kind} (NaN-aware max across sources)"
-            ),
-            "cell_methods": cell_methods,
-            "coordinates": "centroid_lat centroid_lon",
-        }
-    )
-    n_sources.attrs.update(build_n_sources_attrs(n_sources_count))
-
-    ds = xr.Dataset(
-        {
-            "lower_bound": lower,
-            "upper_bound": upper,
-            "n_sources": n_sources,
-        },
-        coords={
-            "time": time_index,
-            id_col: lower[id_col],
-            "time_bnds": time_bnds,
-            "centroid_lat": centroid_lat,
-            "centroid_lon": centroid_lon,
-        },
-    )
-    ds["time"].attrs["bounds"] = "time_bnds"
-    ds["time"].attrs["axis"] = "T"
-    ds["time"].attrs["standard_name"] = "time"
-    ds[id_col].attrs["long_name"] = "HRU identifier"
-    ds[id_col].attrs["cf_role"] = "timeseries_id"
-
-    ds_loaded = ds.compute()
-
-    write_target_nc(
-        ds_loaded,
-        output_path,
-        title=title,
-        extra_global_attrs=extra_global_attrs,
-        sort_dim=id_col,
-    )
-
-    n = ds_loaded["n_sources"].values
-    total = n.size
-    none = int((n == 0).sum())
-    logger.info(
-        "%s coverage: %d/%d cells have >=1 finite source (%.2f%% all-NaN)",
-        bounds_long_name_kind,
-        total - none,
-        total,
-        100.0 * none / total if total else 0.0,
-    )
-
-    if not nn_fill:
-        return
-
-    centroids_xy = hru_meta[["centroid_x", "centroid_y"]].values
-    filled_ds, nn_diag = nn_fill_bounds(
-        ds_loaded, centroids_xy, max_candidates=nn_max_candidates
-    )
-    nn_diag.attrs.update(
-        {
-            "units": "1",
-            "long_name": "nearest-neighbor fill flag",
-            # int8 to match the on-disk nn_filled dtype (CF §3.5).
-            "flag_values": np.array([0, 1], dtype="int8"),
-            "flag_meanings": "not_filled filled",
-            "coordinates": "centroid_lat centroid_lon",
-        }
-    )
-    filled_ds["nn_filled"] = nn_diag
-    filled_attrs = dict(extra_global_attrs)
-    filled_attrs["nn_fill_max_candidates"] = nn_max_candidates
-    filled_attrs["nn_fill_distance_crs"] = project.area_crs
-    nn_path = output_path.with_name(
-        output_path.stem + "_nn_filled" + output_path.suffix
-    )
-    write_target_nc(
-        filled_ds,
-        nn_path,
-        title=nn_title,
-        extra_global_attrs=filled_attrs,
-        sort_dim=id_col,
-    )
-
-
-# ---------------------------------------------------------------------------
-# Year-chunked target build (memory-bounded for multi-year daily targets)
-# ---------------------------------------------------------------------------
-
-
-def iter_period_years(
-    period_start: str, period_end: str
-) -> "list[tuple[int, str, str]]":
-    """Iterate ``(year, year_start_iso, year_end_iso)`` over the period.
-
-    The first and last yields are clipped to ``period_start`` / ``period_end``
-    when those fall mid-year. Used by year-chunked target builders to drive
-    the per-year build loop.
-
-    Examples
-    --------
-    >>> iter_period_years("1980-06-15", "1982-03-20")
-    [(1980, '1980-06-15', '1980-12-31'),
-     (1981, '1981-01-01', '1981-12-31'),
-     (1982, '1982-01-01', '1982-03-20')]
-    """
-    start = pd.Timestamp(period_start)
-    end = pd.Timestamp(period_end)
-    if end < start:
-        raise ValueError(f"period end {period_end!r} precedes start {period_start!r}")
-    out: list[tuple[int, str, str]] = []
-    for year in range(start.year, end.year + 1):
-        ys = max(start, pd.Timestamp(f"{year}-01-01"))
-        ye = min(end, pd.Timestamp(f"{year}-12-31"))
-        out.append((year, ys.strftime("%Y-%m-%d"), ye.strftime("%Y-%m-%d")))
-    return out
-
-
-# ---------------------------------------------------------------------------
-# Cache invalidation for year-chunked intermediates (issue #213)
-# ---------------------------------------------------------------------------
-
-# Names of global attrs written to every per-year intermediate so the skip
-# branch can detect a stale cache. Kept as module-level constants so callers
-# (sca.build, swe.build) can pass the values into ``extra_global_attrs`` and
-# the on-disk schema is canonical.
-INTERMEDIATE_CONFIG_FINGERPRINT_ATTR = "config_fingerprint"
-INTERMEDIATE_CODE_VERSION_ATTR = "code_version"
-
-
-def target_config_fingerprint(project: Project, target_name: str) -> str:
-    """Stable hash of a target's config block + fabric identity.
-
-    Combines the target's config block from ``project.config["targets"]``,
-    the fabric path, and the fabric sha256 into a deterministic
-    sha256-truncated-to-12-hex-chars fingerprint. Used by the year-chunked
-    builders' skip predicate so that any change to the target's knobs
-    (sources, periods, thresholds, etc.) or a fabric swap invalidates the
-    on-disk per-year intermediates automatically.
-
-    The fabric identity is included because a fabric swap changes the HRU
-    set and CRS the target was built against — the same target config
-    against a different fabric produces semantically different output that
-    must not be silently reused.
-
-    Returns
-    -------
-    A 12-character hex string. Collisions across realistic config-edit
-    volumes are negligible.
-    """
-    target_cfg = project.config["targets"][target_name]
-    fabric_cfg = project.config.get("fabric") or {}
-    fabric_sha = (project.fabric or {}).get("sha256", "")
-    payload = {
-        "target": target_cfg,
-        "fabric_path": fabric_cfg.get("path", ""),
-        "fabric_sha256": fabric_sha,
-    }
-    canonical = json.dumps(payload, sort_keys=True, default=str)
-    return hashlib.sha256(canonical.encode("utf-8")).hexdigest()[:12]
-
-
-def code_version_fingerprint() -> str:
-    """The package ``__version__`` used as a code-version cache tag.
-
-    Operators bumping ``__version__`` between runs get automatic
-    invalidation; commits that change builder logic without a version
-    bump do not. Mid-version edits therefore require a manual ``rm``
-    of the intermediates directory — a known limitation documented in
-    each year-chunked builder's module docstring.
-    """
-    from nhf_spatial_targets import __version__
-
-    return __version__
-
-
-def should_skip_year_build(
-    paths: "list[Path]",
-    *,
-    active_config_fingerprint: str,
-    active_code_version: str,
-    target_label: str,
-    year: int,
-    logger: logging.Logger,
-) -> "tuple[bool, dict | None]":
-    """Decide whether to skip rebuilding a year given existing intermediates.
-
-    Replaces the path-existence-only skip predicates that ``targets/sca.py``
-    and ``targets/swe.py`` carried before issue #213. Compares the cached
-    intermediate's ``config_fingerprint`` / ``code_version`` global attrs
-    against the active values; on mismatch (or on missing-attr / corrupt-
-    file), logs a WARNING naming both fingerprints, deletes the stale
-    intermediate(s), and falls through to a rebuild.
-
-    Parameters
-    ----------
-    paths
-        All per-year intermediate paths the caller would expect to find on
-        disk for a complete year build (the unfilled NC plus, when
-        ``nn_fill`` is on, the ``_nn_filled.nc`` companion). The first
-        path is treated as the primary attrs source; companions are read
-        only via ``.exists()``.
-    active_config_fingerprint
-        Output of :func:`target_config_fingerprint` for the current run.
-    active_code_version
-        Output of :func:`code_version_fingerprint` for the current run.
-    target_label
-        Short string used in log messages (e.g. ``"sca"``, ``"swe"``).
-    year
-        Calendar year being built; used in log messages.
-    logger
-        Caller's module logger so warnings appear under the right name.
-
-    Returns
-    -------
-    (skip, cached_attrs)
-        Four distinct return shapes, each with a specific caller contract:
-
-        - ``(False, None)`` — at least one expected path missing. No cache
-          to examine; caller builds the year fresh.
-        - ``(False, {})`` — cached intermediate exists but could not be
-          opened (corrupt / truncated NetCDF). All paths unlinked; caller
-          rebuilds.
-        - ``(False, file_attrs)`` — cached intermediate opened cleanly but
-          its fingerprint attrs were missing (pre-#213 file) or differed
-          from the active values. All paths unlinked; caller rebuilds.
-          ``file_attrs`` is the dict that was read off disk before the
-          unlink, kept for forensic logging.
-        - ``(True, file_attrs)`` — cache is valid; caller may skip the
-          rebuild and use ``file_attrs`` to re-emit any per-year
-          diagnostic warnings (e.g. low-coverage) so the operator sees
-          them on every run, not just the first.
-
-    Callers should defensively guard attr lookups with
-    ``(cached_attrs or {}).get(...)`` to handle the corrupt-file shape
-    without a special case.
-
-    Treats a cached intermediate that is missing the fingerprint attrs
-    as stale (pre-#213 intermediate) and triggers a rebuild — safe
-    default for the one-time upgrade transition. The cleanup pass uses
-    ``unlink(missing_ok=True)`` so a partial intermediate set (unfilled
-    present, _nn_filled missing) does not raise.
-    """
-    if not all(p.exists() for p in paths):
-        return False, None
-
-    primary = paths[0]
-    try:
-        with xr.open_dataset(primary) as cached:
-            cached_attrs = dict(cached.attrs)
-    except (OSError, RuntimeError, ValueError) as exc:
-        # netCDF4-backed xr.open_dataset on a corrupt/truncated file may
-        # raise OSError (HDF5 layer), RuntimeError (libnetcdf bindings),
-        # or ValueError (CF decode failure). All three are "the cache is
-        # unusable" from this function's perspective — recovery is the
-        # same: warn, unlink, let the caller rebuild.
-        logger.warning(
-            "%s year %d: cached intermediate %s could not be opened "
-            "(%s); deleting and rebuilding.",
-            target_label,
-            year,
-            primary,
-            exc,
-        )
-        for p in paths:
-            p.unlink(missing_ok=True)
-        return False, {}
-
-    cached_fp = cached_attrs.get(INTERMEDIATE_CONFIG_FINGERPRINT_ATTR)
-    cached_ver = cached_attrs.get(INTERMEDIATE_CODE_VERSION_ATTR)
-    # ``not`` (rather than ``is None``) also catches the empty-string case,
-    # which a malformed upstream writer could produce — treat as missing
-    # rather than equal-to-active-if-active-is-also-empty.
-    if not cached_fp or not cached_ver:
-        logger.warning(
-            "%s year %d: cached intermediate %s predates issue #213 "
-            "fingerprint support (missing %s/%s attrs); deleting and "
-            "rebuilding.",
-            target_label,
-            year,
-            primary,
-            INTERMEDIATE_CONFIG_FINGERPRINT_ATTR,
-            INTERMEDIATE_CODE_VERSION_ATTR,
-        )
-        for p in paths:
-            p.unlink(missing_ok=True)
-        return False, cached_attrs
-
-    if cached_fp != active_config_fingerprint or cached_ver != active_code_version:
-        logger.warning(
-            "%s year %d: cache fingerprint mismatch — cached config=%s "
-            "code=%s, active config=%s code=%s. Deleting stale "
-            "intermediates and rebuilding.",
-            target_label,
-            year,
-            cached_fp,
-            cached_ver,
-            active_config_fingerprint,
-            active_code_version,
-        )
-        for p in paths:
-            p.unlink(missing_ok=True)
-        return False, cached_attrs
-
-    return True, cached_attrs
-
-
-# ---------------------------------------------------------------------------
-# Orphan pruning for year-chunked intermediates (issue #211)
-# ---------------------------------------------------------------------------
-
-# The 4-digit anchor is load-bearing — a 5-digit "year" must NOT match
-# so that misnamed intermediates fall through to the stitch's
-# computed-from-iter_period_years input list as the second defense layer
-# (verified by test_stitch_input_computed_from_period_not_glob).
-_YEAR_INTERMEDIATE_PATTERN = re.compile(r"_(\d{4})(?:_nn_filled)?\.nc$")
-
-
-def prune_orphan_year_intermediates(
-    intermediates_dir: Path,
-    base: str,
-    expected_years: "set[int]",
-    *,
-    target_label: str,
-    logger: logging.Logger,
-) -> "list[int]":
-    """Delete per-year intermediate NCs whose year falls outside ``expected_years``.
-
-    Globs both ``<base>_<YYYY>.nc`` and ``<base>_<YYYY>_nn_filled.nc``
-    under ``intermediates_dir``, parses the 4-digit year from each
-    filename, and unlinks files whose year is not in ``expected_years``.
-    Mirrors the auto-invalidation pattern from
-    :func:`should_skip_year_build` (#213): the operator sees a WARNING
-    naming the years pruned but does not need to ``rm`` by hand.
-
-    Used by year-chunked target builders (sca, swe) to defend the stitch
-    step against orphan intermediates left behind by a previous build
-    against a wider period (#211). Without this pass the stitcher's
-    directory glob would silently include them, leaking out-of-period
-    years into the final canonical NC. Files matching the glob but
-    failing the strict year regex (e.g. operator-renamed debris) are
-    logged at DEBUG and left untouched.
-
-    Parameters
-    ----------
-    intermediates_dir
-        ``<project>/targets/.<target>_intermediates/`` for the current
-        build. Must exist (caller's responsibility).
-    base
-        Filename prefix matching the per-year writer's pattern, without
-        the trailing ``_<YYYY>``. E.g. ``"sca_targets"`` matches
-        ``sca_targets_2005.nc`` and ``sca_targets_2005_nn_filled.nc``.
-    expected_years
-        Set of calendar years the active config period covers. Files
-        whose parsed year is NOT in this set are pruned.
-    target_label
-        Short string used in log messages (e.g. ``"sca"``, ``"swe"``).
-    logger
-        Caller's module logger so the WARNING appears under the right
-        name.
-
-    Returns
-    -------
-    Sorted list of years pruned. Empty when no orphans were found.
-    """
-    pruned: dict[int, list[Path]] = {}
-    skipped: list[str] = []
-    for path in intermediates_dir.glob(f"{base}_*.nc"):
-        match = _YEAR_INTERMEDIATE_PATTERN.search(path.name)
-        if not match:
-            # Filename didn't match the canonical pattern — skip rather
-            # than risk unlinking an unrelated file. Recorded for the
-            # DEBUG breadcrumb so operators auditing the intermediates
-            # dir know prune saw them and chose not to act.
-            skipped.append(path.name)
-            continue
-        year = int(match.group(1))
-        if year in expected_years:
-            continue
-        pruned.setdefault(year, []).append(path)
-
-    if skipped:
-        logger.debug(
-            "%s: %d files in %s matched the glob but not the year "
-            "regex; left untouched (names: %s)",
-            target_label,
-            len(skipped),
-            intermediates_dir,
-            ", ".join(sorted(skipped)),
-        )
-
-    if not pruned:
-        # Audit breadcrumb — operators tailing the build log can confirm
-        # the prune pass actually ran when nothing was found.
-        logger.debug(
-            "%s: no orphan year intermediates to prune (active period covers %d years)",
-            target_label,
-            len(expected_years),
-        )
-        return []
-
-    for year_paths in pruned.values():
-        for path in year_paths:
-            path.unlink(missing_ok=True)
-
-    pruned_years = sorted(pruned)
-    logger.warning(
-        "%s: pruned %d orphan year intermediates outside the active "
-        "period (years: %s). Caused by a downward period change since "
-        "the previous build; without pruning the stitched output would "
-        "silently include these years. This is automatic recovery; no "
-        "action required.",
-        target_label,
-        sum(len(v) for v in pruned.values()),
-        ", ".join(str(y) for y in pruned_years),
-    )
-    return pruned_years
-
-
-def stitch_year_chunks_to_target(
-    intermediate_files: "list[Path]",
-    output_path: Path,
-    *,
-    title: str,
-    extra_global_attrs: dict | None,
-    sort_dim: str,
-) -> None:
-    """Lazily open per-year target NCs and stream them into one canonical NC.
-
-    Streams as **full-time HRU columns**: the dataset is rechunked to
-    ``{"time": -1, sort_dim: chunk_hru}`` so a single dask chunk spans the
-    whole time axis for a block of HRUs. This matches the columnar on-disk
-    layout that :func:`io_nc.build_encoding` writes (``(full_time, chunk_hru)``)
-    — each on-disk chunk is filled in one pass, never re-compressed across year
-    boundaries (issue #165 ST3b). Peak memory is one ~256 MiB column block, not
-    the whole target.
-
-    (This replaces the earlier year-slab streaming, which bounded memory the
-    same way but produced a time-slab on-disk layout that made per-HRU
-    calibration reads pull every time chunk.)
-
-    Encoding is delegated to :func:`io_nc.build_encoding` (``layer="target"``):
-    float32+zlib bounds, int8+zlib diagnostics, pinned ``proleptic_gregorian``
-    time. Global attrs are taken from the first per-year file, then overridden
-    by ``extra_global_attrs`` plus a fresh ``history`` line. The output is
-    sorted ascending on ``sort_dim`` (issue #93) and written atomically via
-    :func:`io_nc.atomic_to_netcdf`.
-
-    Parameters
-    ----------
-    intermediate_files
-        Per-year NC paths to stitch (already sorted by year via filename glob).
-    output_path
-        Final canonical NC path (e.g. ``<project>/targets/swe_targets.nc``).
-    title
-        Global ``title`` attr for the stitched file.
-    extra_global_attrs
-        Per-target metadata to overlay on the per-year files' global attrs.
-    sort_dim
-        Dimension to sort ascending on before write (typically
-        ``project.id_col``); also the HRU chunk dimension.
-    """
-    from datetime import datetime, timezone
-
-    from nhf_spatial_targets import __version__
-    from nhf_spatial_targets.io_nc import atomic_to_netcdf, build_encoding
-
-    if not intermediate_files:
-        raise ValueError(
-            "stitch_year_chunks_to_target: intermediate_files is empty; "
-            "the year loop produced no per-year NCs."
-        )
-
-    output_path = Path(output_path)
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-
-    ds = xr.open_mfdataset(
-        [str(p) for p in intermediate_files],
-        combine="by_coords",
-        # `join="exact"` mirrors the project's defensive-failure pattern
-        # in `check_hru_coords`: a coord-set mismatch across per-year
-        # files indicates per-year-build corruption (fabric drift,
-        # weight-cache poisoning) and should fail loud, not silently
-        # broadcast NaN on the union as `join="outer"` would.
-        join="exact",
-        chunks={},  # native per-file chunks; rechunked to columns below
-        engine="netcdf4",
-    )
-    try:
-        ds = ds.sortby(sort_dim)
-        # Rechunk to full-time HRU columns so the columnar on-disk write fills
-        # each chunk in one pass (no cross-year re-compression). chunk_hru here
-        # (~256 MiB read block) is intentionally wider than the ~1 MiB on-disk
-        # chunk: it bounds streaming memory while still spanning the full time
-        # axis, so every on-disk chunk it covers is written exactly once.
-        n_time = int(ds.sizes.get("time", 1))
-        ds = ds.chunk({"time": -1, sort_dim: _read_chunk_hru(n_time, 4)})
-
-        # `combine_attrs="override"` (the open_mfdataset default) keeps
-        # the first file's attrs, which include per-year-only attrs
-        # like ``year_chunk`` and ``frac_valid_bound`` (SCA's per-year
-        # low-coverage diagnostic) that don't apply to the stitched
-        # multi-year output. Strip them before the canonical attrs are
-        # applied so the final NC reports its actual scope honestly —
-        # otherwise an operator inspecting the stitched NC's attrs
-        # would trust a single-year statistic as if it were
-        # multi-year.
-        for per_year_attr in ("year_chunk", "frac_valid_bound"):
-            ds.attrs.pop(per_year_attr, None)
-        ds.attrs.setdefault("Conventions", "CF-1.6")
-        ds.attrs["title"] = title
-        ds.attrs["history"] = (
-            f"{datetime.now(timezone.utc).isoformat()} stitched from "
-            f"{len(intermediate_files)} per-year NCs by "
-            f"nhf_spatial_targets v{__version__}"
-        )
-        ds.attrs.setdefault("institution", "USGS")
-        ds.attrs.setdefault("software_version", __version__)
-        if extra_global_attrs:
-            ds.attrs.update(extra_global_attrs)
-
-        target_dtypes = {
-            v: "float32" for v in ("lower_bound", "upper_bound") if v in ds.data_vars
-        }
-        target_dtypes.update(
-            {v: "int8" for v in ("n_sources", "nn_filled") if v in ds.data_vars}
-        )
-        encoding = build_encoding(
-            ds,
-            layer="target",
-            hru_dim=sort_dim,
-            var_dtype=target_dtypes,
-            timesteps_per_file=n_time,
-        )
-        expected_sizes = {"time": n_time, sort_dim: int(ds.sizes[sort_dim])}
-        atomic_to_netcdf(ds, output_path, encoding=encoding)
-    finally:
-        ds.close()
-
-    # Cheap post-write coverage check: a truncated or mis-shaped stream would
-    # change the dim sizes. The per-year intermediates remain on disk, so a
-    # loud failure here is recoverable (re-run regenerates the stitch).
-    with xr.open_dataset(output_path) as _check:
-        got_sizes = {
-            "time": _check.sizes.get("time"),
-            sort_dim: _check.sizes.get(sort_dim),
-        }
-    if got_sizes != expected_sizes:
-        raise RuntimeError(
-            f"stitched output {output_path} has dims {got_sizes}, expected "
-            f"{expected_sizes}; the stitch did not write the full extent."
-        )
-
-    logger.info(
-        "Stitched %d per-year NCs -> %s (%.1f MB)",
-        len(intermediate_files),
-        output_path,
-        output_path.stat().st_size / 1e6,
-    )
+from nhf_spatial_targets.targets._combine import (
+    build_n_sources_attrs,
+    multi_source_nanminmax,
+)
+from nhf_spatial_targets.targets._intermediates import (
+    INTERMEDIATE_CODE_VERSION_ATTR,
+    INTERMEDIATE_CONFIG_FINGERPRINT_ATTR,
+    code_version_fingerprint,
+    iter_period_years,
+    prune_orphan_year_intermediates,
+    should_skip_year_build,
+    stitch_year_chunks_to_target,
+    target_config_fingerprint,
+)
+from nhf_spatial_targets.targets._io import (
+    _load_and_reproject_fabric,
+    _read_chunk_hru,
+    check_hru_coords,
+    compute_hru_area_and_centroids,
+    compute_hru_areas,
+    compute_hru_centroids,
+    parse_period,
+    read_aggregated_source,
+    reindex_to_day_start,
+    reindex_to_month_start,
+)
+from nhf_spatial_targets.targets._shims import (
+    SourceShim,
+    shims_by_config_label,
+    shims_by_key,
+    validate_source_units,
+)
+from nhf_spatial_targets.targets._writers import (
+    _target_encoding_without_chunks,
+    write_bounds_target,
+    write_target_nc,
+)
+
+__all__ = [
+    # _shims
+    "SourceShim",
+    "shims_by_config_label",
+    "shims_by_key",
+    "validate_source_units",
+    # _io
+    "_read_chunk_hru",
+    "_load_and_reproject_fabric",
+    "check_hru_coords",
+    "compute_hru_area_and_centroids",
+    "compute_hru_areas",
+    "compute_hru_centroids",
+    "parse_period",
+    "read_aggregated_source",
+    "reindex_to_day_start",
+    "reindex_to_month_start",
+    # _combine
+    "build_n_sources_attrs",
+    "multi_source_nanminmax",
+    # _writers
+    "_target_encoding_without_chunks",
+    "write_bounds_target",
+    "write_target_nc",
+    # _intermediates
+    "INTERMEDIATE_CODE_VERSION_ATTR",
+    "INTERMEDIATE_CONFIG_FINGERPRINT_ATTR",
+    "code_version_fingerprint",
+    "iter_period_years",
+    "prune_orphan_year_intermediates",
+    "should_skip_year_build",
+    "stitch_year_chunks_to_target",
+    "target_config_fingerprint",
+]
